notebooks/requirements-build.txt
anywidget>=0.9.0,<0.10.0
google-api-python-client>=2,<3
google-auth>=2,<3
google-auth-oauthlib>=1,<2
ipykernel>=6,<7
jupyter>=1,<2
matplotlib>=3,<4
nbformat>=5,<6
numpy>=2,<3
onnx>=1,<2
pycairo>=1,<2
tqdm>=4,<5
